services/multi/requirements.txt
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.3
colorama==0.4.6
Flask==2.3.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
py-zipkin==1.2.8
python-dotenv==1.0.0
requests==2.31.0
typing_extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.3.6
